#50 adiciona transformação de objeto
